add some conversion stuff to rtl

+++ volkrnl/src/rtl/rtlmem.c
@@ -3,8 +3,8 @@
 // slower functions, will optimize later on
 
 void* rtl_copy_mem(const void* src, void* dest, u64 len){
-    const i64* srcInt = src;
-    i64* destInt = dest;
+    const i8* srcInt = src;
+    i8* destInt = dest;
     for(u64 i = 0; i < len; i++){
         destInt[i] = srcInt[i];
     }
@@ -12,7 +12,7 @@ void* rtl_copy_mem(const void* src, void* dest, u64 len){
 }
 
 void* rtl_set_mem(void* dest, u64 value, u64 len){
-    i64* destInt = dest;
+    i8* destInt = dest;
 
     for(u64 i = 0; i < len; i++){
         destInt[i] = value;
@@ -25,9 +25,9 @@ void rtl_zero_mem(void* dest, u64 len){
 }
 
 i64 rtl_compare_mem(const void* first, const void* second, u64 len){
-    const i64* firstInt = first;
-    const i64* secondInt = second;
-    for(int i = 0; i < len; i++){
+    const i8* firstInt = first;
+    const i8* secondInt = second;
+    for(u64 i = 0; i < len; i++){
         if(firstInt[i] != secondInt[i]){
             return firstInt[i] - secondInt[i];
         }
